+++ client/tests/unit/widgets/Header.ts
@@ -34,6 +34,12 @@ describe('Header', () => {
 					w(Link, { to: 'home', classes: ['navbar-brand'] }, ['Block Lang']),
 					v('ul', { classes: ['navbar-nav ml-auto'] }, [
 						v('li', { classes: ['nav-item'] }, [
+							w(Link, { to: '/projects/new', classes: ['nav-link'] }, [
+								v('i', { classes: ['fas fa-plus'] }, []),
+								' 创建项目'
+							])
+						]),
+						v('li', { classes: ['nav-item'] }, [
 							w(Link, { to: '/user/jack', classes: ['nav-link'] }, [
 								v('img', { classes: [css.avatar], src: '#', width: 20, height: 20 }, []),
 								' jack'